Transform guild.rs into its own mod
